better zoom segment dragging
